==> 2.gender.rmd <==
---
title: "2. Gender"
author: "by: Nina Branten"
bibliography: references.bib
---



```{r, install remotely, globalsettings, echo=FALSE, warning=FALSE, results='hide', eval=FALSE}
install.packages("remotes")
remotes::install_github("rlesur/klippy")
``` 

```{r, globalsettings, echo=FALSE, warning=FALSE, results='hide'}
library(knitr)
library(rgl)

knitr::opts_chunk$set(echo = TRUE)
opts_chunk$set(tidy.opts=list(width.cutoff=100),tidy=TRUE, warning = FALSE, message = FALSE,comment = "#>", cache=TRUE, class.source=c("test"), class.output=c("test2"))
options(width = 100)
rgl::setupKnitr()



colorize <- function(x, color) {sprintf("<span style='color: %s;'>%s</span>", color, x) }

```

```{r klippy, echo=FALSE, include=TRUE}
klippy::klippy(position = c('top', 'right'))
#klippy::klippy(color = 'darkred')
#klippy::klippy(tooltip_message = 'Click to copy', tooltip_success = 'Done')
```


# getting started

```{r, eval=FALSE}
#start with clean workspace 
rm(list=ls())
getwd()
load("./data/names_df_v20220106.RData")
```

# packages

```{r}
library(data.table) 
library(tidyverse) 
require(xml2)
require(rvest)
require(devtools)
require(scholar)
require(stringi)

```



## adding gender to dataset




### gender to cs_df
```{r, eval=FALSE}
names_df %>% mutate(firstname=first_name) -> names_df
names_df$gender <- NA

# make links to scrape
names_df$name_url <- paste0("https://www.meertens.knaw.nl/nvb/naam/is/", names_df[, c("firstname")])

name_list <- list()
table_list <- list()

for (i in 1:nrow(names_df)) {
    print(i)
    if (!(is.na(names_df$gender[i]))) next
    name_list[[i]] <- read_html(names_df$name_url[i])
    # extract name frequency table and gender info
    if (length(name_list[[i]] %>% html_table())>0) {
      
      table_list[[i]] <- name_list[[i]] %>% html_table()
      table_list[[i]][[1]][table_list[[i]][[1]]=="--"] <- "0"
      if (as.numeric(table_list[[i]][[1]]$X3[2]) > as.numeric(table_list[[i]][[1]]$X3[6])) {
        names_df$gender[i] <- "male" } else {
          names_df$gender[i] <- "female"
        }
    }
    save(names_df, file="./data/names_df_temp.RData")
}


```

```{r, eval=FALSE}
#devtools::install_github("kalimu/genderizeR")
require(genderizeR)
names_df$gender2 <- NA

for (i in 1:nrow(names_df)) {
    print(i)
    tryCatch({
      if (!(is.na(names_df$gender2[i]))) next
      gender <- genderizeAPI(names_df$first_name[i])
      gender <- gender[[1]]
      names_df$gender2[i] <- gender$gender[which.max(gender$probability)]
    }, 
    warning = function(w) {
        cat("WARNING:", conditionMessage(w), "\n") #WARNING message
    },
    error=function(e){
      cat("Error:", conditionMessage(e), "\n") #ERROR message
    })  
    save(names_df, file="./data/names_df_temp.RData")
}




```


```{r, eval=FALSE}
save(names_df, file="./data/names_df_v20221005.RData")
```
